+++ src/components/Loading.js
@@ -19,13 +19,13 @@ const Round = styled.div`
     animation: lds-ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
     border-color: #fff transparent transparent transparent;
   }
-  div:nth-child(1) {
+  div:nth-of-type(1) {
     animation-delay: -0.45s;
   }
-  div:nth-child(2) {
+  div:nth-of-type(2) {
     animation-delay: -0.3s;
   }
-  div:nth-child(3) {
+  div:nth-of-type(3) {
     animation-delay: -0.15s;
   }
   @keyframes lds-ring {
